+++ rust/operator-binary/src/druid_connection_controller.rs
@@ -31,9 +31,9 @@ use crate::{
     crd::{
         INTERNAL_SECRET_SECRET_KEY, PYTHONPATH, SUPERSET_CONFIG_FILENAME, druidconnection, v1alpha1,
     },
-    rbac,
+    operations::job_state::{JobState, get_job_state},
+    resources::rbac,
     superset_controller::CONTAINER_IMAGE_BASE_NAME,
-    util::{JobState, get_job_state},
 };
 
 pub const DRUID_CONNECTION_CONTROLLER_NAME: &str = "druid-connection";