+++ rust/operator-binary/src/main.rs
@@ -12,7 +12,7 @@ use stackable_operator::{
     crd::authentication::core,
     eos::EndOfSupportChecker,
     k8s_openapi::api::{
-        apps::v1::StatefulSet,
+        apps::v1::{Deployment, StatefulSet},
         batch::v1::Job,
         core::v1::{ConfigMap, Service},
     },
@@ -44,18 +44,12 @@ use crate::{
 };
 
 mod authorization;
-mod commands;
 mod config;
-mod controller_commons;
 mod crd;
 mod druid_connection_controller;
-mod listener;
 mod operations;
-mod product_logging;
-mod rbac;
-mod service;
+mod resources;
 mod superset_controller;
-mod util;
 mod webhooks;
 
 mod built_info {
@@ -156,6 +150,11 @@ async fn main() -> anyhow::Result<()> {
                 )
                 .owns(
                     watch_namespace.get_api::<DeserializeGuard<StatefulSet>>(&client),
+                    watcher::Config::default(),
+                )
+                // Required for workers and beat.
+                .owns(
+                    watch_namespace.get_api::<DeserializeGuard<Deployment>>(&client),
                     watcher::Config::default(),
                 )
                 .watches(